@@ -22,37 +22,36 @@
         "@emotion/react": "^11.8.2",
         "@emotion/styled": "^11.8.1",
         "@mui/icons-material": "^5.5.1",
-        "@mui/material": "^5.5.1",
+        "@mui/material": "^5.5.3",
         "@reduxjs/toolkit": "^1.7.2",
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
+        "react": "18.0.0",
+        "react-dom": "18.0.0",
         "react-redux": "^7.2.6",
         "react-router-dom": "^6.2.2"
     },
     "devDependencies": {
-        "@babel/core": "^7.17.7",
+        "@babel/core": "^7.17.8",
         "@babel/plugin-transform-runtime": "^7.17.0",
         "@babel/preset-env": "^7.16.11",
         "@babel/preset-react": "^7.16.7",
         "@babel/preset-typescript": "^7.16.7",
-        "@babel/runtime": "^7.17.7",
-        "@typechain/ethers-v5": "^9.0.0",
-        "@types/react": "^17.0.40",
-        "@types/react-dom": "^17.0.13",
+        "@typechain/ethers-v5": "^10.0.0",
+        "@types/react": "^17.0.43",
+        "@types/react-dom": "^17.0.14",
         "@types/react-router-dom": "^5.3.3",
-        "@typescript-eslint/eslint-plugin": "^5.15.0",
-        "@typescript-eslint/parser": "^5.15.0",
-        "babel-loader": "^8.2.3",
+        "@typescript-eslint/eslint-plugin": "^5.17.0",
+        "@typescript-eslint/parser": "^5.17.0",
+        "babel-loader": "^8.2.4",
         "copy-webpack-plugin": "^10.2.4",
         "css-loader": "^6.7.1",
-        "eslint": "^8.11.0",
+        "eslint": "^8.12.0",
         "eslint-plugin-react": "^7.29.4",
-        "eslint-plugin-react-hooks": "^4.3.0",
+        "eslint-plugin-react-hooks": "^4.4.0",
         "eslint-webpack-plugin": "^3.1.1",
         "html-webpack-plugin": "^5.5.0",
         "style-loader": "^3.3.1",
-        "typechain": "^7.0.1",
-        "typescript": "^4.6.2",
+        "typechain": "^8.0.0",
+        "typescript": "^4.6.3",
         "webpack": "^5.70.0",
         "webpack-cli": "^4.9.2",
         "webpack-dev-server": "^4.7.4"
