@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
 import { store } from "./redux/store";
 import { Provider } from "react-redux";
@@ -24,4 +24,4 @@ function App() {
     );
 }
 
-ReactDOM.render(<App />, document.getElementById("react-container"));
+createRoot(document.getElementById("react-container")).render(<App />);
